Make deviceorientationabsoluteevent WPT test match spec
The deviceorientationabsoluteevent WPT test currently asserts that the
'absolute' property of the deviceorientationabsolute event is false.

The spec states that:
https://w3c.github.io/deviceorientation/spec-source-orientation.html#deviceorientationabsolute
"The deviceorientationabsolute event is completely analogous to the
deviceorientation event, except additional sensors like the magnetometer can be
used to provide an absolute orientation. The absolute property must be set to
true."

This updates the test to assert that absolute is set to true.

Bug: 774183, 802815
Change-Id: I3694e2b774aa0e4a8a9f7d0328e2aad8b37775e6