[css-layout-api] Adds ability to set available sizes on children.
This adds the availableInlineSize, availableBlockSize, to the layout
constriants options dictionary.

This also makes the children "shrink-to-fit" if they are inside a
custom layout pass. Additional logic was needed to *not* "shrink-to-fit"
children during the layout fallback if any occurred, so
LayoutCustomPhaseScope was added to handle this.

The tests that were added test a new BFC with inline children to
see if shrink to fitting works as expected.

Bug: 726125
Change-Id: Ic86d51579b680d871c66040f54e532bc47d79979